+++ packages/components/src/components/layouts/Flex/Flex.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Meta } from '@storybook/react';
 import { Flex } from './Flex';
-import { Box, Text } from 'components';
+import { Box, Text } from '@components';
 
 export default {
   component: Flex,
@@ -21,12 +21,7 @@ export default {
 const Template = ({ children, ...rest }) => (
   <Flex {...rest}>
     {[1, 2, 3, 4, 5].map((index) => (
-      <Box
-        key={index}
-        margin="spacer-1"
-        padding="spacer-4"
-        border="1px solid var(--chakra-colors-accent-default)"
-      >
+      <Box key={index} margin="spacer-1" padding="spacer-4" border="1px solid var(--chakra-colors-accent-default)">
         <Text variant="text-body-bold">
           {children}:{` ${index}`}
         </Text>