+++ packages/components/src/components/layouts/Flex/Flex.test.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { screen, renderWithProviders } from 'tests/renderWithProviders';
+import { screen, renderWithProviders } from '@tests/renderWithProviders';
 import { Flex } from './Flex';
 
 describe('Flex', () => {
-  const givenComponentRendered = () =>
-    renderWithProviders(<Flex>Children</Flex>);
+  const givenComponentRendered = () => renderWithProviders(<Flex>Children</Flex>);
 
   test('should render a children', () => {
     givenComponentRendered();