+++ packages/components/src/components/primitives/Select/Select.tsx
@@ -1,26 +1,31 @@
 import React, { FC } from 'react';
 import { Select as ChakraSelect, useMultiStyleConfig } from '@chakra-ui/react';
 import { SelectProps } from './types';
+import { IconChevronDown } from '..';
 
 export const Select: FC<SelectProps> = ({ value, placeholder, ...props }) => {
   const {
-    field: { color, placeholderColor },
+    field: {
+      color,
+      placeholderColor,
+      _disabled: { disabledPlaceHolderColor },
+    },
   } = useMultiStyleConfig('Select', { value, iconSize: props.iconSize }) as {
-    field: { color: string; placeholderColor: string };
+    field: { color: string; placeholderColor: string; _disabled: { disabledPlaceHolderColor: string } };
   };
 
   let textColor = color;
 
   if (placeholder && (!value || value === '')) {
-    textColor = placeholderColor;
+    textColor = props?.isDisabled ? disabledPlaceHolderColor : placeholderColor;
   }
 
   return (
     <ChakraSelect
       data-testid="cmpsr.select"
-      color={textColor}
       textColor={textColor}
       placeholder={placeholder}
+      icon={<IconChevronDown />}
       {...props}
     />
   );