+++ packages/components/src/components/primitives/Switch/Switch.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { fireEvent, screen } from '@testing-library/react';
-import { renderWithProviders } from 'tests/renderWithProviders';
+import { renderWithProviders } from '@tests/renderWithProviders';
 import { Switch } from '.';
 import { useMultiStyleConfig } from '@chakra-ui/react';
 
@@ -35,9 +35,7 @@ describe('Switch', () => {
   });
 
   test('it should return proper styling when checked', () => {
-    const { container } = renderWithProviders(
-      <Switch labelPosition="left" label="test" />
-    );
+    const { container } = renderWithProviders(<Switch labelPosition="left" label="test" />);
     const input = container.querySelector('input');
     fireEvent.click(input);
     expect(useMultiStyleConfig).toHaveBeenCalledWith('Switch', {
@@ -47,9 +45,7 @@ describe('Switch', () => {
 
   test('it should handle the onChange event', () => {
     const onChange = jest.fn();
-    const { container } = renderWithProviders(
-      <Switch labelPosition="left" label="test" onChange={onChange} />
-    );
+    const { container } = renderWithProviders(<Switch labelPosition="left" label="test" onChange={onChange} />);
     const input = container.querySelector('input');
     fireEvent.click(input);
     expect(onChange).toHaveBeenCalled();